v2.0

+++ backend/CreamyBunApp/function/universalFunctions.py
@@ -33,8 +33,16 @@ def get_now_time():
 # 删除举报任务信息及对应图片
 def delete_a_reported_task_all(report_id):
     r = get_a_reported_task(report_id)
-    os.remove(r.image_url)
+    if r.image_url != 'image' and r.image_url != "":
+        os.remove(r.image_url)
     delete_a_reported_task(report_id)
+
+# 检验jwt
+def check_jwt(username,jwt):
+    u = get_a_user_data(username)
+    if u.jwt == "" or u.jwt != jwt:
+        return False
+    return True
 
 # 发送邮件并返回验证码
 def send_email(email):
@@ -200,7 +208,7 @@ def get_task_info_list(username, state, page_number, sort_choice):
 def change_user_avatar(image, username):
     avatar_url = USER_AVATAR_SAVE_PATH + username + "." + image.name
     update_avatar_url_by_username(username, avatar_url)
-    with open(avatar_url, 'wb') as f:
+    with open(avatar_url, 'wb+') as f:
         for line in image:
             f.write(line)
 
@@ -227,7 +235,10 @@ def set_admin_password(new_password):
 
 # 获取用户头像base64格式
 def get_user_avatr(username):
-    avatar_url = get_a_user_data(username).avatar_url
+    u = get_a_user_data(username)
+    if not u:
+        return ""
+    avatar_url = u.avatar_url
     if not os.path.exists(avatar_url):
         return ""
     else:
@@ -298,8 +309,8 @@ def walk_file(file, material_type):
     j = 1
     output_dirs = []
     for base, dirs, _ in os.walk(file):
+        dirs.sort()
         for d in dirs:
-
             dirPath = os.path.join(base, d)
             sub_list = []
             i = 1
@@ -308,6 +319,7 @@ def walk_file(file, material_type):
                 # dirs 表示该文件夹下的子目录名list
                 # files 表示该文件夹下的文件list
                 # 遍历文件
+                files.sort()
                 for f in files:
                     file_type = os.path.splitext(f)[-1][1:]
                     file_path = os.path.join(root, f)
@@ -563,6 +575,10 @@ def user_receive_current_task(username,task_id):
     if u.credit_rank < t.star_rank:
         return False, 'lowRank'
 
+    # 没题目可领了就不领，防止前端多调用本接口出错
+    if len(t.receiver_list.all()) == t.max_receiver_number:
+        return False, 'noProblemLeft'
+
     p_list = t.problem_list.all()
 
     # 确定测试题列表（包括资质检测和穿插题目）
@@ -580,10 +596,17 @@ def user_receive_current_task(username,task_id):
     td = TaskDict.objects.create(task_id=task_id,task_status_for_user=HAS_RECEIVED,\
                                 task_status_for_itself=NOT_FINISHED,test_problem_number=before_test_number)
 
+    # 确定当前用户应该领多少题
+    problem_number_to_receive = t.problem_number_for_single_receiver
+    if t.left_problem_number > 0:
+        now_receive_problem_number = len([p for p in p_list if p.current_state != NOT_RECEIVED])
+        if now_receive_problem_number - len(t.receiver_list.all()) * problem_number_to_receive < t.left_problem_number:
+            problem_number_to_receive += 1
+
     # 要做的题目列表
     normal_test_list = []
     for p in p_list:
-        if len(normal_test_list) >= t.problem_number_for_single_receiver:
+        if len(normal_test_list) >= problem_number_to_receive:
             break
         if p.is_test == False and p.current_state == NOT_RECEIVED:
             normal_test_list.append(p)
@@ -818,5 +841,5 @@ def force_release_task(task_id):
     set_task_begin_time(t,now_time)
     get_task_status(t)
     sub_donut_number = t.single_bonus*t.problem_total_number
-    flag = sub_donut_for_user(u,sub_donut_number)
+    flag = True
     return flag, sub_donut_number, u.donut_number
